Chat / DevTools: simplify after the connector-UI rework
Three round-trip-style cleanups on top of this session's UI changes:

* Extract `MessageAction` into its own module. Was inline in Chat.tsx;
  the DevTools panel rewrite replicated the same `cursor / opacity 0.7
  → 1 on hover` style, and the new helper-text inline actions tripled
  it. Pulling the component out lets both consumers reuse it via
  `className` overrides for color / weight / margin, and avoids a
  Chat ↔ DevTools import cycle.

* Severity / gating predicates now route through the existing
  `Connectors.stateRequiresAck` and `Connectors.stateIsTransient`.
  `helperSeverityFor` and `showAck` were inlining the same checks; the
  predicates have been the source-of-truth since qhq-5d0.2 and now
  this layer agrees with the lifecycle layer.

* Drop the `<>{helperLines}</>` Fragment, the
  `(x): x is React.ReactElement` type guard, and the verbose comment
  blocks that mostly restated the code. Pluralize `Ready — N tool(s)`
  in the DevTools panel so single-tool connectors don't say "1 tools".
